pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pycaching"
version = "0.1.0"
description = "Modular Python caching library for microservices, databases, and LLM workflows"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Kiran Roy"}
]
keywords = ["caching", "cache", "redis", "memcached", "llm", "performance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
]

dependencies = [
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
redis = [
    "redis>=5.0.0",
]
memcached = [
    "pymemcache>=4.0.0",
]
mongodb = [
    "pymongo>=4.0.0",
]
aws = [
    "boto3>=1.28.0",
]
cloudflare = [
    "cloudflare>=2.0.0",
]
llm = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
visualization = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "matplotlib>=3.7.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
]
config = [
    "pyyaml>=6.0",
]
serialization = [
    "msgpack>=1.0.0",
    "orjson>=3.9.0",
]
all = [
    "pycaching[redis,memcached,mongodb,aws,cloudflare,llm,visualization,config,serialization]",
]

[project.urls]
Homepage = "https://github.com/RoyKiran/pycaching"
Documentation = "https://github.com/RoyKiran/pycaching#readme"
Repository = "https://github.com/RoyKiran/pycaching"
Issues = "https://github.com/RoyKiran/pycaching/issues"

[tool.setuptools]
packages = ["pycaching"]

[tool.setuptools.package-data]
pycaching = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
